Se crea una subcarpeta en Controllers llamada Admin y se mueven las clases  ImageController y ProductController aca por lo que se actualizan las rutas en el archivo web.php y se actualiza el namespace de cada archivo y ademas se importa la clase Controller con use.

+++ routes/web.php
@@ -22,16 +22,16 @@ Route::get('/home', 'HomeController@index')->name('home');
 Route::middleware(['auth', 'admin'])->prefix('admin')->group(function () {
 
     // Administracion de productos
-    Route::get('/products', 'ProductController@index');
-    Route::get('/products/create', 'ProductController@create');
-    Route::post('/products', 'ProductController@store');
-    Route::get('/products/{id}/edit', 'ProductController@edit');
-    Route::post('/products/{id}/edit', 'ProductController@update');
-    Route::delete('/products/{id}/delete', 'ProductController@destroy');
+    Route::get('/products', 'Admin\ProductController@index');
+    Route::get('/products/create', 'Admin\ProductController@create');
+    Route::post('/products', 'Admin\ProductController@store');
+    Route::get('/products/{id}/edit', 'Admin\ProductController@edit');
+    Route::post('/products/{id}/edit', 'Admin\ProductController@update');
+    Route::delete('/products/{id}/delete', 'Admin\ProductController@destroy');
 
-    Route::get('/products/{id}/images', 'ImageController@index');
-    Route::post('/products/{id}/images', 'ImageController@store');
-    Route::delete('/products/{id}/images', 'ImageController@destroy');
-    Route::get('/products/{id}/images/select/{image}', 'ImageController@select'); // Destacar imagen
+    Route::get('/products/{id}/images', 'Admin\ImageController@index');
+    Route::post('/products/{id}/images', 'Admin\ImageController@store');
+    Route::delete('/products/{id}/images', 'Admin\ImageController@destroy');
+    Route::get('/products/{id}/images/select/{image}', 'Admin\ImageController@select'); // Destacar imagen
 
 });
